requirements.txt
Django==5.2.8
Pillow==10.1.0
qrcode==7.4.2
python-decouple==3.8
djangorestframework==3.15.2
django-cors-headers==4.5.0
psycopg2-binary==2.9.11
redis==5.2.0
sentry-sdk==2.19.0
python-json-logger==2.0.7
whitenoise==6.8.2
gunicorn==23.0.0
